+++ lib/ts/recipe/thirdparty/api/signinup.ts
@@ -16,6 +16,7 @@
 import STError from "../error";
 import { send200Response } from "../../../utils";
 import { APIInterface, APIOptions } from "../";
+import { findRightProvider } from "../utils";
 
 export default async function signInUpAPI(apiImplementation: APIInterface, options: APIOptions): Promise<boolean> {
     if (apiImplementation.signInUpPOST === undefined) {
@@ -27,6 +28,7 @@ export default async function signInUpAPI(apiImplementation: APIInterface, optio
     let code = bodyParams.code === undefined ? "" : bodyParams.code;
     let redirectURI = bodyParams.redirectURI;
     let authCodeResponse = bodyParams.authCodeResponse;
+    let clientId = bodyParams.clientId;
 
     if (thirdPartyId === undefined || typeof thirdPartyId !== "string") {
         throw new STError({
@@ -63,18 +65,32 @@ export default async function signInUpAPI(apiImplementation: APIInterface, optio
         });
     }
 
-    let provider = options.providers.find((p) => p.id === thirdPartyId);
+    let provider = findRightProvider(options.providers, thirdPartyId, clientId);
     if (provider === undefined) {
-        throw new STError({
-            type: STError.BAD_INPUT_ERROR,
-            message:
-                "The third party provider " +
-                thirdPartyId +
-                " seems to not be configured on the backend. Please check your frontend and backend configs.",
-        });
+        if (clientId === undefined) {
+            throw new STError({
+                type: STError.BAD_INPUT_ERROR,
+                message: "The third party provider " + thirdPartyId + ` seems to be missing from the backend configs.`,
+            });
+        } else {
+            throw new STError({
+                type: STError.BAD_INPUT_ERROR,
+                message:
+                    "The third party provider " +
+                    thirdPartyId +
+                    ` seems to be missing from the backend configs. If it is configured, then please make sure that you are passing the correct clientId from the frontend.`,
+            });
+        }
     }
 
-    let result = await apiImplementation.signInUpPOST({ provider, code, redirectURI, options, authCodeResponse });
+    let result = await apiImplementation.signInUpPOST({
+        provider,
+        code,
+        clientId,
+        redirectURI,
+        options,
+        authCodeResponse,
+    });
 
     if (result.status === "OK") {
         send200Response(options.res, {