+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -270,6 +270,7 @@ export type APIInterface = {
               code: string;
               redirectURI: string;
               authCodeResponse?: any;
+              clientId?: string;
               options: ThirdPartyAPIOptions;
           }) => Promise<
               | {
@@ -322,4 +323,8 @@ export type APIInterface = {
                     status: "EMAIL_ALREADY_EXISTS_ERROR";
                 }
           >);
+
+    appleRedirectHandlerPOST:
+        | undefined
+        | ((input: { code: string; state: string; options: ThirdPartyAPIOptions }) => Promise<void>);
 };