+++ CHANGELOG.md
@@ -10,7 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Support for FDI 1.10: Allow thirdparty `/signinup POST` API to take `authCodeResponse` XOR `code` so that it can supprt OAuth via PKCE
+-   Support for FDI 1.10:
+    -   Allow thirdparty `/signinup POST` API to take `authCodeResponse` XOR `code` so that it can supprt OAuth via PKCE
+    -   Adds apple sign in callback API
+-   Optional `getRedirectURI` function added to social providers in case we set the `redirect_uri` on the backend.
+-   Adds optional `isDefault` param to auth providers so that they can be reused with different credentials.
 
 ## [8.0.4] - 2021-10-28